requirements.txt
absl-py==1.1.0
anyio==3.6.1
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
fastapi==0.78.0
fastapi-csrf-protect==0.2.2
filelock==3.7.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h11==0.13.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
librosa==0.9.2
llvmlite==0.38.1
Markdown==3.3.7
MarkupSafe==2.1.1
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.3.0
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
SoundFile==0.10.3.post1
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.2
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
watchfiles==0.15.0
